+++ packages/ipfs-cli/src/commands/key/list.js
@@ -1,6 +1,9 @@
 'use strict'
 
 const parseDuration = require('parse-duration').default
+const {
+  stripControlCharacters
+} = require('../../utils')
 
 module.exports = {
   command: 'list',
@@ -18,6 +21,6 @@ module.exports = {
     const keys = await ipfs.key.list({
       timeout
     })
-    keys.forEach((ki) => print(`${ki.id} ${ki.name}`))
+    keys.forEach((ki) => print(`${ki.id} ${stripControlCharacters(ki.name)}`))
   }
 }