+++ packages/ipfs-cli/test/files/ls.js
@@ -99,10 +99,24 @@ describe('ls', () => {
     expect(output).to.include(files[0].size)
   })
 
-  it('should list a path with details', async () => {
+  it('should list a path with a timeout', async () => {
+    const path = '/foo'
+
+    await cli(`files ls ${path} --timeout=1s`, { ipfs, print })
+
+    expect(ipfs.files.ls.callCount).to.equal(1)
+    expect(ipfs.files.ls.getCall(0).args).to.deep.equal([
+      path, {
+        ...defaultOptions,
+        timeout: 1000
+      }
+    ])
+  })
+
+  it('should strip control characters from path names', async () => {
     const files = [{
       cid: fileCid,
-      name: 'file-name',
+      name: 'file\n\t\b-name',
       size: 'file-size',
       mode: 0o755,
       mtime: {
@@ -117,43 +131,7 @@ describe('ls', () => {
 
     expect(ipfs.files.ls.callCount).to.equal(1)
     expect(output).to.include(files[0].cid.toString())
-    expect(output).to.include(files[0].name)
+    expect(output).to.include('file-name')
     expect(output).to.include(files[0].size)
-  })
-
-  it('should list a path with details (short option)', async () => {
-    const files = [{
-      cid: fileCid,
-      name: 'file-name',
-      size: 'file-size',
-      mode: 0o755,
-      mtime: {
-        secs: Date.now() / 1000,
-        nsecs: 0
-      }
-    }]
-
-    ipfs.files.ls = sinon.stub().withArgs('/foo', defaultOptions).returns(files)
-
-    await cli('files ls -l /foo', { ipfs, print })
-
-    expect(ipfs.files.ls.callCount).to.equal(1)
-    expect(output).to.include(files[0].cid.toString())
-    expect(output).to.include(files[0].name)
-    expect(output).to.include(files[0].size)
-  })
-
-  it('should list a path with a timeout', async () => {
-    const path = '/foo'
-
-    await cli(`files ls ${path} --timeout=1s`, { ipfs, print })
-
-    expect(ipfs.files.ls.callCount).to.equal(1)
-    expect(ipfs.files.ls.getCall(0).args).to.deep.equal([
-      path, {
-        ...defaultOptions,
-        timeout: 1000
-      }
-    ])
   })
 })