+++ packages/ipfs-core-utils/src/files/normalise-input/normalise-input.js
@@ -8,7 +8,9 @@ const {
   isBytes,
   isBlob,
   isReadableStream,
-  isFileObject
+  isFileObject,
+  mtimeToObject,
+  modeToNumber
 } = require('./utils')
 
 // eslint-disable-next-line complexity
@@ -46,12 +48,15 @@ module.exports = async function * normaliseInput (input, normaliseContent) {
 
   // Iterable<?>
   if (input[Symbol.iterator] || input[Symbol.asyncIterator]) {
+    /** @type {any} peekable */
     const peekable = itPeekable(input)
+
+    /** @type {any} value **/
     const { value, done } = await peekable.peek()
 
     if (done) {
       // make sure empty iterators result in empty files
-      yield * peekable
+      yield * []
       return
     }
 
@@ -103,7 +108,8 @@ async function toFileObject (input, normaliseContent) {
   // @ts-ignore - Those properties don't exist on most input types
   const { path, mode, mtime, content } = input
 
-  const file = { path: path || '', mode, mtime }
+  const file = { path: path || '', mode: modeToNumber(mode), mtime: mtimeToObject(mtime) }
+
   if (content) {
     file.content = await normaliseContent(content)
   } else if (!path) { // Not already a file object with path or content prop