+++ packages/ipfs-core-utils/src/index.js
@@ -1,1 +1,7 @@
 'use strict'
+
+/**
+ * @template {any[]} ARGS
+ * @template R
+ * @typedef {(...args: ARGS) => R} Fn
+ */