+++ packages/ipfs-core/src/components/add.js
@@ -25,7 +25,7 @@ module.exports = ({ addAll }) => {
  * @property {string} [hashAlg] - multihash hashing algorithm to use (default: `'sha2-256'`)
  * @property {boolean} [onlyHash] - If true, will not add blocks to the blockstore (default: `false`)
  * @property {boolean} [pin] - pin this object when adding (default: `true`)
- * @property {Function} [progress] - a function that will be called with the byte length of chunks as a file is added to ipfs (default: `undefined`)
+ * @property {(bytes:number, path:string) => void} [progress] - a function that will be called with the number of bytes added as a file is added to ipfs and the path of the file being added
  * @property {boolean} [rawLeaves] - if true, DAG leaves will contain raw file data and not be wrapped in a protobuf (default: `false`)
  * @property {boolean} [trickle] - if true will use the [trickle DAG](https://godoc.org/github.com/ipsn/go-ipfs/gxlibs/github.com/ipfs/go-unixfs/importer/trickle) format for DAG generation (default: `false`)
  * @property {boolean} [wrapWithDirectory] - Adds a wrapping node around the content (default: `false`)