+++ packages/ipfs-core/src/components/name/pubsub/subs.js
@@ -1,7 +1,7 @@
 'use strict'
 
 const { getPubsubRouting } = require('./utils')
-const { withTimeoutOption } = require('../../../utils')
+const withTimeoutOption = require('ipfs-core-utils/src/with-timeout-option')
 
 module.exports = ({ ipns, options: constructorOptions }) => {
   /**