+++ packages/ipfs-core/src/components/object/stat.js
@@ -1,7 +1,7 @@
 'use strict'
 
 const dagPB = require('ipld-dag-pb')
-const { withTimeoutOption } = require('../../utils')
+const withTimeoutOption = require('ipfs-core-utils/src/with-timeout-option')
 
 module.exports = ({ ipld, preload }) => {
   const get = require('./get')({ ipld, preload })