+++ packages/ipfs-core/src/components/start.js
@@ -11,7 +11,7 @@ const routingConfig = require('../ipns/routing/config')
 const { AlreadyInitializedError, NotEnabledError } = require('../errors')
 const Components = require('./')
 const createMfsPreload = require('../mfs-preload')
-const { withTimeoutOption } = require('../utils')
+const withTimeoutOption = require('ipfs-core-utils/src/with-timeout-option')
 
 const WEBSOCKET_STAR_PROTO_CODE = 479