+++ packages/ipfs-message-port-client/.aegir.js
@@ -4,7 +4,7 @@ const EchoServer = require('aegir/utils/echo-server')
 const echoServer = new EchoServer()
 
 module.exports = {
-  bundlesize: { maxSize: '12kB' },
+  bundlesize: { maxSize: '13kB' },
   karma: {
     files: [
       {