+++ packages/ipfs-http-gateway/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.3](https://github.com/ipfs/js-ipfs/compare/ipfs-http-gateway@0.1.2...ipfs-http-gateway@0.1.3) (2020-11-25)
+
+**Note:** Version bump only for package ipfs-http-gateway
+
+
+
+
+
+## [0.1.2](https://github.com/ipfs/js-ipfs/compare/ipfs-http-gateway@0.1.1...ipfs-http-gateway@0.1.2) (2020-11-16)
+
+**Note:** Version bump only for package ipfs-http-gateway
+
+
+
+
+
+## [0.1.1](https://github.com/ipfs/js-ipfs/compare/ipfs-http-gateway@0.1.0...ipfs-http-gateway@0.1.1) (2020-11-09)
+
+**Note:** Version bump only for package ipfs-http-gateway
+
+
+
+
+
 # 0.1.0 (2020-10-28)